fix: use mcp icons in tool sidebar header
